Rename to CircuitProvider and CircuitContent
